feat: Sidebar で完結するようになった
